group call/discover jitsi: send discover_jitsi request on button click
